# trvCollection.postman_collection.json-Swagger20.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "trvCollection",
    "description": "# Introduction\nSimple express API with 5 main functionalities.\n * Managing rowing days appointments.\n * User authentification.\n * Managing news of the website.\n * Managing user data.\n * Enabling pdf download.\n",
    "contact": {}
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/addNews": {
      "post": {
        "description": "Add a News",
        "summary": "http://localhost:3000/api/addNews",
        "tags": [
          "Misc"
        ],
        "operationId": "http://localhost:3000/api/addNews",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-auth-token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/http%3A~1~1localhost%3A3000~1api~1addNewsRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/http%3A~1~1localhost%3A3000~1api~1addNews"
            },
            "examples": {
              "application/json": {
                "successMsg": "News created!",
                "news": {
                  "id": 11,
                  "title": "Ruderkurs für Anfänger 2020, 8 Teilnehmer waren begeistert!",
                  "preview_text": "Sechs Übungseinheiten wurden absolviert und die Rookies wurden für den Einstieg in den Rudersport ausgebildet. Der",
                  "body": "Sechs Übungseinheiten wurden",
                  "created_at": null,
                  "image_url": "trv_news/gzspy1nybqfrvvmo2m2l"
                }
              }
            },
            "headers": {
              "x-powered-by": {
                "type": "string",
                "default": "Express"
              },
              "access-control-allow-origin": {
                "type": "string",
                "default": "*"
              },
              "content-type": {
                "type": "string",
                "default": "application/json; charset=utf-8"
              },
              "content-length": {
                "type": "string",
                "default": "354"
              },
              "etag": {
                "type": "string",
                "default": "W/\"162-+06OHhmh0J6GyibOp2CF7ssUj34\""
              },
              "date": {
                "type": "string",
                "default": "Tue, 08 Dec 2020 20:16:55 GMT"
              },
              "connection": {
                "type": "string",
                "default": "close"
              },
              "Vary": {
                "type": "string",
                "default": "Accept-Encoding"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/http%3A~1~1localhost%3A3000~1api~1users"
            },
            "examples": {
              "application/json": {
                "msg": "Please enter all fields!"
              }
            }
          }
        }
      }
    },
...
    "/news/delete/{id}": {
      "delete": {
        "description": "Delete a specific news",
        "summary": "http://localhost:3000/api/news/delete/:id",
        "tags": [
          "Misc"
        ],
        "operationId": "http://localhost:3000/api/news/delete/:id",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/http%3A~1~1localhost%3A3000~1api~1news~1delete~1%3Aid"
            },
            "examples": {
              "application/json": {
                "news": [
                  {
                    "id": 11,
                    "title": "Ruderkurs für Anfänger 2020, 8 Teilnehmer waren begeistert!",
                    "preview_text": "Sechs Übungseinheiten wurden absolviert und die Rookies wurden für den Einstieg in den Rudersport ausgebildet. Der",
                    "body": "Sechs Übungseinheiten wurden",
                    "created_at": null,
                    "image_url": "trv_news/gzspy1nybqfrvvmo2m2l"
                  },
                  {
                    "id": 10,
                    "title": "sXAXAXAxaXAXAXXAxaxxaxaxaxaxax",
                    "preview_text": "ads",
                    "body": "adsa",
                    "created_at": null,
                    "image_url": "trv_news/r24q7zltkehp2ijibfee"
                  },
                  {
                    "id": 8,
                    "title": "Ruderkurs für Anfänger 2020, 8 Teilnehmer waren begeistert!",
                    "preview_text": "Sechs Übungseinheiten wurden absolviert und die Rookies wurden für den Einstieg in den Rudersport ausgebildet. Der nächste Kurs kommt bestimmt. Ab vier Anmeldungen starten wir wieder Ende August. Interessenten melden sich bei Karl S. Nolz - Mobil 0660-2030660",
                    "body": "Sechs Übungseinheiten wurden absolviert und die Rookies wurden für den Einstieg in den Rudersport ausgebildet. Der nächste Kurs kommt bestimmt. Ab vier Anmeldungen starten wir wieder Ende August. Interessenten melden sich bei Karl S. Nolz - Mobil 0660-2030660Sechs Übungseinheiten wurden absolviert und die Rookies wurden für den Einstieg in den Rudersport ausgebildet. Der nächste Kurs kommt bestimmt. Ab vier Anmeldungen starten wir wieder Ende August. Interessenten melden sich bei Karl S. Nolz - Mobil 0660-2030660Sechs Übungseinheiten wurden absolviert und die Rookies wurden für den Einstieg in den Rudersport ausgebildet. Der nächste Kurs kommt bestimmt. Ab vier Anmeldungen starten wir wieder Ende August. Interessenten melden sich bei Karl S. Nolz - Mobil 0660-2030660Sechs Übungseinheiten wurden absolviert und die Rookies wurden für den Einstieg in den Rudersport ausgebildet. Der nächste Kurs kommt bestimmt. Ab vier Anmeldungen starten wir wieder Ende August. Interessenten melden sich bei Karl S. Nolz - Mobil 0660-2030660Sechs Übungseinheiten wurden absolviert und die Rookies wurden für den Einstieg in den Rudersport ausgebildet. Der nächste Kurs kommt bestimmt. Ab vier Anmeldungen starten wir wieder Ende August. Interessenten melden sich bei Karl S. Nolz - Mobil 0660-2030660",
                    "created_at": null,
                    "image_url": "trv_news/azwszqvl99prmftblrai"
                  },
                  {
                    "id": 7,
                    "title": "Obmannwechsel beim TRV an Gerald Schaffer",
                    "preview_text": "Herbert Hiesinger übergibt bei der Hauptversammlung am 27. Juni 2020 und blickt auf ereignisreiche Jahre zurück.",
                    "body": "Herbert Hiesinger übergibt bei der Hauptversammlung am 27. Juni 2020 und blickt auf ereignisreiche Jahre zurück.Herbert Hiesinger übergibt bei der Hauptversammlung am 27. Juni 2020 und blickt auf ereignisreiche Jahre zurück.Herbert Hiesinger übergibt bei der Hauptversammlung am 27. Juni 2020 und blickt auf ereignisreiche Jahre zurück.Herbert Hiesinger übergibt bei der Hauptversammlung am 27. Juni 2020 und blickt auf ereignisreiche Jahre zurück.Herbert Hiesinger übergibt bei der Hauptversammlung am 27. Juni 2020 und blickt auf ereignisreiche Jahre zurück.Herbert Hiesinger übergibt bei der Hauptversammlung am 27. Juni 2020 und blickt auf ereignisreiche Jahre zurück.Herbert Hiesinger übergibt bei der Hauptversammlung am 27. Juni 2020 und blickt auf ereignisreiche Jahre zurück.Herbert Hiesinger übergibt bei der Hauptversammlung am 27. Juni 2020 und blickt auf ereignisreiche Jahre zurück.Herbert Hiesinger übergibt bei der Hauptversammlung am 27. Juni 2020 und blickt auf ereignisreiche Jahre zurück.",
                    "created_at": null,
                    "image_url": "trv_news/jx8k9eobkidfbusgov9y"
                  },
                  {
                    "id": 6,
                    "title": "JAHRESBERICHT des TRV 2019",
                    "preview_text": "Der TRV kann auf eine ereignisreiche Saison zurückblicken.  Die aktiven Mitglieder ruderderten 28.000 km. Jahresbericht 2019 als PDF (6,6 MB) 2019 sind wieder zahlreiche Aktivitäten geplant -> sh. Terminkalender 2020 FOTO: Ruderkurs für Anfänger",
                    "body": "Der TRV kann auf eine ereignisreiche Saison zurückblicken. \nDie aktiven Mitglieder ruderderten 28.000 km. Jahresbericht 2019 als PDF (6,6 MB)\n2019 sind wieder zahlreiche Aktivitäten geplant -> sh. Terminkalender 2020\nFOTO: Ruderkurs für AnfängerDer TRV kann auf eine ereignisreiche Saison zurückblicken. \nDie aktiven Mitglieder ruderderten 28.000 km. Jahresbericht 2019 als PDF (6,6 MB)\n2019 sind wieder zahlreiche Aktivitäten geplant -> sh. Terminkalender 2020\nFOTO: Ruderkurs für AnfängerDer TRV kann auf eine ereignisreiche Saison zurückblicken. \nDie aktiven Mitglieder ruderderten 28.000 km. Jahresbericht 2019 als PDF (6,6 MB)\n2019 sind wieder zahlreiche Aktivitäten geplant -> sh. Terminkalender 2020\nFOTO: Ruderkurs für AnfängerDer TRV kann auf eine ereignisreiche Saison zurückblicken. \nDie aktiven Mitglieder ruderderten 28.000 km. Jahresbericht 2019 als PDF (6,6 MB)\n2019 sind wieder zahlreiche Aktivitäten geplant -> sh. Terminkalender 2020\nFOTO: Ruderkurs für Anfänger",
                    "created_at": null,
                    "image_url": "trv_news/wd7raxf4sm1xdstybcyo"
                  },
                  {
                    "id": 4,
                    "title": "Österr. Ruderindoor Meisterschaften,",
                    "preview_text": "Gratulation an die TRV-Ruderer für 4 Podestplätze:  GOLD für Team mixed, 2000m - Johanna Hiesinger, Doris Pennetzdorfer, Max Brönimann, Leonhard Riemer.  SILBER Männer, 500m Leonhard Riemer BRONZE Männer Masters C, 1000m – Reinhard Tauber BRONZE Team Masters mixed C, 2000m – Marianne Riemer, Doris Pennetzdorfer, Reinhard Tauber, Franz Gratsch",
                    "body": "Gratulation an die TRV-Ruderer für 4 Podestplätze: \nGOLD für Team mixed, 2000m - Johanna Hiesinger, Doris Pennetzdorfer, Max Brönimann, Leonhard Riemer. \nSILBER Männer, 500m Leonhard Riemer\nBRONZE Männer Masters C, 1000m – Reinhard Tauber\nBRONZE Team Masters mixed C, 2000m – Marianne Riemer, Doris Pennetzdorfer, Reinhard Tauber, Franz GratschGratulation an die TRV-Ruderer für 4 Podestplätze: \nGOLD für Team mixed, 2000m - Johanna Hiesinger, Doris Pennetzdorfer, Max Brönimann, Leonhard Riemer. \nSILBER Männer, 500m Leonhard Riemer\nBRONZE Männer Masters C, 1000m – Reinhard Tauber\nBRONZE Team Masters mixed C, 2000m – Marianne Riemer, Doris Pennetzdorfer, Reinhard Tauber, Franz GratschGratulation an die TRV-Ruderer für 4 Podestplätze: \nGOLD für Team mixed, 2000m - Johanna Hiesinger, Doris Pennetzdorfer, Max Brönimann, Leonhard Riemer. \nSILBER Männer, 500m Leonhard Riemer\nBRONZE Männer Masters C, 1000m – Reinhard Tauber\nBRONZE Team Masters mixed C, 2000m – Marianne Riemer, Doris Pennetzdorfer, Reinhard Tauber, Franz GratschGratulation an die TRV-Ruderer für 4 Podestplätze: \nGOLD für Team mixed, 2000m - Johanna Hiesinger, Doris Pennetzdorfer, Max Brönimann, Leonhard Riemer. \nSILBER Männer, 500m Leonhard Riemer\nBRONZE Männer Masters C, 1000m – Reinhard Tauber\nBRONZE Team Masters mixed C, 2000m – Marianne Riemer, Doris Pennetzdorfer, Reinhard Tauber, Franz Gratsch",
                    "created_at": null,
                    "image_url": "trv_news/c2shbhachdoia3ecev1c"
                  },
                  {
                    "id": 3,
                    "title": "TRV-Schlumpfachter beim Vienna NightRow",
                    "preview_text": "Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  ",
                    "body": "Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  Lorem ipsum. blablabla.  ",
                    "created_at": null,
                    "image_url": "trv_news/jyogfhkxtgjvah68tzbg"
                  },
                  {
                    "id": 2,
                    "title": "Tullner Erfolg am Mondsee",
                    "preview_text": "Bei der 10 km Langstreckenregatta am Mondsee erreichte unser Mixed Doppelzweier mit Anne Mück und Franz Gratsch mit 46,18 Min. die 9. Zeit von 27 Startern. Sie gewannen damit ihre Klasse MXC klar.",
                    "body": "Bei der 10 km Langstreckenregatta am Mondsee erreichte unser Mixed Doppelzweier mit Anne Mück und Franz Gratsch mit 46,18 Min. die 9. Zeit von 27 Startern. Sie gewannen damit ihre Klasse MXC klar.Bei der 10 km Langstreckenregatta am Mondsee erreichte unser Mixed Doppelzweier mit Anne Mück und Franz Gratsch mit 46,18 Min. die 9. Zeit von 27 Startern. Sie gewannen damit ihre Klasse MXC klar.Bei der 10 km Langstreckenregatta am Mondsee erreichte unser Mixed Doppelzweier mit Anne Mück und Franz Gratsch mit 46,18 Min. die 9. Zeit von 27 Startern. Sie gewannen damit ihre Klasse MXC klar.Bei der 10 km Langstreckenregatta am Mondsee erreichte unser Mixed Doppelzweier mit Anne Mück und Franz Gratsch mit 46,18 Min. die 9. Zeit von 27 Startern. Sie gewannen damit ihre Klasse MXC klar.Bei der 10 km Langstreckenregatta am Mondsee erreichte unser Mixed Doppelzweier mit Anne Mück und Franz Gratsch mit 46,18 Min. die 9. Zeit von 27 Startern. Sie gewannen damit ihre Klasse MXC klar.",
                    "created_at": null,
                    "image_url": "trv_news/vuaehklluqlfusdlr1e5"
                  }
                ],
                "successMsg": "News was deleted!"
              }
            },
            "headers": {
              "x-powered-by": {
                "type": "string",
                "default": "Express"
              },
              "access-control-allow-origin": {
                "type": "string",
                "default": "*"
              },
              "content-type": {
                "type": "string",
                "default": "application/json; charset=utf-8"
              },
              "etag": {
                "type": "string",
                "default": "W/\"2606-FWbRSIw/ppjHD3+kkctwQrFKlwE\""
              },
              "date": {
                "type": "string",
                "default": "Tue, 08 Dec 2020 20:21:24 GMT"
              },
              "connection": {
                "type": "string",
                "default": "close"
              },
              "Vary": {
                "type": "string",
                "default": "Accept-Encoding"
              },
              "Content-Encoding": {
                "type": "string",
                "default": "gzip"
              },
              "Transfer-Encoding": {
                "type": "string",
                "default": "chunked"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/http%3A~1~1localhost%3A3000~1api~1users"
            },
            "examples": {
              "application/json": {
                "msg": "News do not exist!"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/http%3A~1~1localhost%3A3000~1api~1users"
            },
            "examples": {
              "application/json": {
                "msg": "No token, authorization denied!"
              }
            }
          }
        }
      }
    },
    "/news/{id}": {
      "get": {
        "description": "Get a specific news",
        "summary": "http://localhost:3000/api/news/:id",
        "tags": [
          "Misc"
        ],
        "operationId": "http://localhost:3000/api/news/:id",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/http%3A~1~1localhost%3A3000~1api~1news~1%3Aid"
              }
            },
            "examples": {
              "application/json": [
                {
                  "id": 4,
                  "title": "Österr. Ruderindoor Meisterschaften,",
                  "preview_text": "Gratulation an die TRV-Ruderer für 4 Podestplätze:  GOLD für Team mixed, 2000m - Johanna Hiesinger, Doris Pennetzdorfer, Max Brönimann, Leonhard Riemer.  SILBER Männer, 500m Leonhard Riemer BRONZE Männer Masters C, 1000m – Reinhard Tauber BRONZE Team Masters mixed C, 2000m – Marianne Riemer, Doris Pennetzdorfer, Reinhard Tauber, Franz Gratsch",
                  "body": "Gratulation an die TRV-Ruderer für 4 Podestplätze: \nGOLD für Team mixed, 2000m - Johanna Hiesinger, Doris Pennetzdorfer, Max Brönimann, Leonhard Riemer. \nSILBER Männer, 500m Leonhard Riemer\nBRONZE Männer Masters C, 1000m – Reinhard Tauber\nBRONZE Team Masters mixed C, 2000m – Marianne Riemer, Doris Pennetzdorfer, Reinhard Tauber, Franz GratschGratulation an die TRV-Ruderer für 4 Podestplätze: \nGOLD für Team mixed, 2000m - Johanna Hiesinger, Doris Pennetzdorfer, Max Brönimann, Leonhard Riemer. \nSILBER Männer, 500m Leonhard Riemer\nBRONZE Männer Masters C, 1000m – Reinhard Tauber\nBRONZE Team Masters mixed C, 2000m – Marianne Riemer, Doris Pennetzdorfer, Reinhard Tauber, Franz GratschGratulation an die TRV-Ruderer für 4 Podestplätze: \nGOLD für Team mixed, 2000m - Johanna Hiesinger, Doris Pennetzdorfer, Max Brönimann, Leonhard Riemer. \nSILBER Männer, 500m Leonhard Riemer\nBRONZE Männer Masters C, 1000m – Reinhard Tauber\nBRONZE Team Masters mixed C, 2000m – Marianne Riemer, Doris Pennetzdorfer, Reinhard Tauber, Franz GratschGratulation an die TRV-Ruderer für 4 Podestplätze: \nGOLD für Team mixed, 2000m - Johanna Hiesinger, Doris Pennetzdorfer, Max Brönimann, Leonhard Riemer. \nSILBER Männer, 500m Leonhard Riemer\nBRONZE Männer Masters C, 1000m – Reinhard Tauber\nBRONZE Team Masters mixed C, 2000m – Marianne Riemer, Doris Pennetzdorfer, Reinhard Tauber, Franz Gratsch",
                  "created_at": null,
                  "image_url": "trv_news/c2shbhachdoia3ecev1c"
                }
              ]
            },
            "headers": {
              "x-powered-by": {
                "type": "string",
                "default": "Express"
              },
              "access-control-allow-origin": {
                "type": "string",
                "default": "*"
              },
              "content-type": {
                "type": "string",
                "default": "application/json; charset=utf-8"
              },
              "etag": {
                "type": "string",
                "default": "W/\"78e-iruShUKwR37YgF6aCM5U40zIqbM\""
              },
              "date": {
                "type": "string",
                "default": "Tue, 08 Dec 2020 20:22:55 GMT"
              },
              "connection": {
                "type": "string",
                "default": "close"
              },
              "Vary": {
                "type": "string",
                "default": "Accept-Encoding"
              },
              "Content-Encoding": {
                "type": "string",
                "default": "gzip"
              },
              "Transfer-Encoding": {
                "type": "string",
                "default": "chunked"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/http%3A~1~1localhost%3A3000~1api~1users"
            },
            "examples": {
              "application/json": {
                "msg": "News do not exist!"
              }
            }
          }
        }
      }
    },
...
  },
...
}